refactor(support-hub): fetch tenant-scoped collections over the adapter seam
Browser TanStack DB collections now read `/api/admin/support/**` instead of
hardcoded Give Hope seed. Schema lives in support-hub.schema.ts. Messages stay
local-only identity; privileged writes remain server-command owned.
