+++ apps/admin/features/support-hub/components/settings/automations/AutomationRuleForm.tsx
@@ -71,7 +71,7 @@ export function AutomationRuleForm({
   const previewRule: SupportAutomationRule = React.useMemo(
     () => ({
       id: rule?.id ?? "draft",
-      tenantId: rule?.tenantId ?? "tenant-give-hope",
+      tenantId: rule?.tenantId ?? "draft",
       name: name || "Untitled rule",
       description: description ? description : null,
       enabled,
