+++ apps/admin/features/support-hub/stores/support-store.ts
@@ -32,10 +32,11 @@ import {
 import type { SupportReplyPayload } from "../models/editor-payload";
 
 /**
- * Single namespaced surface for everything the hooks layer consumes. By
- * routing through `supportStore.collections.*` and `supportStore.inputs.*`,
- * a later phase can swap the in-memory collection writers for
- * `@asym/api/admin/support-hub/*` mutations without touching any hook file.
+ * Single namespaced surface for everything the hooks layer consumes.
+ * `supportStore.collections` is the TanStack DB read surface over the
+ * adapter-backed `/api/admin/support/**` routes. The messages collection is
+ * local-only identity; thread messages stay on `useSupportMessages`.
+ * Privileged writes stay server-command owned in `use-support-mutations.ts`.
  *
  * Adapter rule: do NOT call collection methods (insert/update/delete) from
  * outside the hooks module. Hooks own the optimistic flow + invalidation;
